Add Makefile, improve build_apptainer_images.sh and run.sh

+++ run.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 dset_dir="$1"
+algorithm="$2" 
 spectra_dir="$dset_dir/mgf"
 output_root_dir="./outputs"
 overlay_size=1024
@@ -8,7 +9,14 @@ overlay_size=1024
 dset_name=$(basename "$dset_dir")
 output_dir="$output_root_dir/$dset_name"
 
-recalculate=false
+# Echo message based on whether an algorithm is provided
+if [ -z "$algorithm" ]; then
+    echo "Running benchmark with all algorithms on datset $dset_name."
+else
+    echo "Running benchmark with $algorithm on dataset $dset_name."
+fi
+
+recalculate=true; # false;
 
 while getopts ":r" opt; do
   case $opt in
@@ -37,39 +45,46 @@ for algorithm_dir in algorithms/*; do
     # if [ -d "$algorithm_dir" ]; then
     if [ -d "$algorithm_dir" ] && [ $(basename "$algorithm_dir") != "base" ]; then
         algorithm_name=$(basename "$algorithm_dir")
-        output_file="$output_dir/${algorithm_name}_output.csv"
-        echo "Output file: $output_file"
-        
-        # Check if the output file does not exist
-        if [ ! -e "$output_file" ]; then
-            echo "Processing algorithm: $algorithm_name"
 
-            # Remove an existing container overlay, if any
-            rm -rf "algorithms/${algorithm_name}/overlay.img"
-            # Create writable overlay for the container
-            apptainer overlay create --fakeroot --size $overlay_size --sparse "algorithms/${algorithm_name}/overlay.img"
+        # If an algorithm is specified, only continue if algorithm_name matches
+        if [ -z "$algorithm" ] || [ "$algorithm_name" == "$algorithm" ]; then
 
-            # Calculate predictions
-            echo "RUN ALGORITHM"
-            apptainer exec --fakeroot --nv \
-                --overlay "algorithms/${algorithm_name}/overlay.img" \
-                -B "${spectra_dir}":"/algo/${dset_name}" \
-                --env-file .env \
-                "algorithms/${algorithm_name}/container.sif" \
-                bash -c "cd /algo && ./make_predictions.sh ${dset_name}"
+            output_file="$output_dir/${algorithm_name}_output.csv"
+            echo "Output file: $output_file"
             
-            # Collect predictions in output_dir
-            echo "EXPORT PREDICTIONS"
-            apptainer exec --fakeroot \
-                --overlay "algorithms/${algorithm_name}/overlay.img" \
-                -B "${output_dir}":/algo/outputs \
-                --env-file .env \
-                "algorithms/${algorithm_name}/container.sif" \
-                bash -c "cp /algo/outputs.csv /algo/outputs/${algorithm_name}_output.csv"
+            # Check if the output file does not exist
+            if [ ! -e "$output_file" ]; then
+                echo "Processing algorithm: $algorithm_name"
 
-        else
-            echo "Skipping algorithm: $algorithm_name. Output file already exists."
+                # Remove an existing container overlay, if any
+                rm -rf "algorithms/${algorithm_name}/overlay.img"
+                # Create writable overlay for the container
+                apptainer overlay create --fakeroot --size $overlay_size --sparse "algorithms/${algorithm_name}/overlay.img"
+
+                # Calculate predictions
+                echo "RUN ALGORITHM $algorithm_name"
+                apptainer exec --fakeroot --nv \
+                    --overlay "algorithms/${algorithm_name}/overlay.img" \
+                    -B "${spectra_dir}":"/algo/${dset_name}" \
+                    --env-file .env \
+                    "algorithms/${algorithm_name}/container.sif" \
+                    bash -c "cd /algo && ./make_predictions.sh ${dset_name}"
+                
+                # Collect predictions in output_dir
+                echo "EXPORT PREDICTIONS"
+                apptainer exec --fakeroot \
+                    --overlay "algorithms/${algorithm_name}/overlay.img" \
+                    -B "${output_dir}":/algo/outputs \
+                    --env-file .env \
+                    "algorithms/${algorithm_name}/container.sif" \
+                    bash -c "cp /algo/outputs.csv /algo/outputs/${algorithm_name}_output.csv"
+
+            else
+                echo "Skipping algorithm: $algorithm_name. Output file already exists."
+            fi
+
         fi
+
     fi
 done
